Big Changed
asd

+++ HowlOut/Models/SearchSettings.cs
@@ -6,17 +6,20 @@ namespace HowlOut
 	public class SearchSettings
 	{
 		public List<EventType> EventTypes { get; set; }
+		//public List<int> EventTypesAsNumbers { get; set; }
 
-		public int Distance { get; set; }
+		public double Distance { get; set; }
+
 
 		public DateTime StartDate { get; set; }
 		public DateTime EndDate { get; set; }
-
+		/*
 		public DateTime StartTime { get; set; }
 		public DateTime EndTime { get; set; }
-
+*/
 		public SearchSettings ()
 		{
+			
 		}
 	}
 }
